pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Personal AI assistant powered by local models"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3,<9",
    "rich>=14.0,<15",
    "httpx>=0.28,<1",
    "openai>=2.28,<3",
    "PyYAML>=6.0,<7",
    "chromadb>=1.5,<2",
    "sentence-transformers>=5.3,<6",
    "feedparser>=6.0,<7",
    "icalendar>=7.0,<8",
    "PyPDF2>=3.0,<4",
    "ddgs>=9.0,<10",
    "pyttsx3>=2.90,<3",
]

[project.optional-dependencies]
voice = ["openai-whisper>=20230918"]

[project.scripts]
jarvis = "jarvis.cli:main"

[tool.setuptools.packages.find]
include = ["jarvis*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
